requirements.txt
streamlit==1.35.0
pymupdf
openpyxl
python-docx
st-pages
python-dotenv